==> 4c.MigConnectivity_WIFL.Rmd <==
---
title: "MigConnectivity-WIFL"
author: "Christen Bossu"
date: "6/14/2024"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

## Install packages for MigConnectivity to estimate strength of connectivity and mignette for calculating abundances of breeding and nonbreeding nodes and visualization of transition probabilities

```
#new version on CRAN has NMC calculation
install.packages("MigConnectivity")
# install.packages("remotes")
remotes::install_github("mgdesaix/mignette")

```

```{r}
Sys.setenv("PROJ_LIB"="/Library/Frameworks/R.framework/Resources/library/sf/proj")
library(MigConnectivity)
library(mignette)
library(terra)
library(raster)
library(tidyverse)
library(sf)
library(lwgeom)
```

## Read in polygons of breeding genoscape. Can first convert genoscape raster to polygons in mignette

```{r}
wifl_genoscape<-terra::rast("input/MigConn/WIFL//WIFL_Genoscape_Spat_Raster_K4.tif")

genoscape_polygons <- mignette::scape_to_shape(x = wifl_genoscape, 
                                               prob_threshold = 0.5)


genoscape_polygons2<- genoscape_polygons %>% mutate(br_node=if_else(Cluster=="WIFL_Genoscape_Spat_Raster_K4_1","EST",if_else(Cluster=="WIFL_Genoscape_Spat_Raster_K4_2","SD",if_else(Cluster=="WIFL_Genoscape_Spat_Raster_K4_3","SW","PNW_INW"))))

##colorblind colors plus navy
cluster_colors <-  c(
  INW = "#CC79A7", #F0E442yellow, interior west
  EST = "#56B4E9", #cyan blue, east
  PNW_INW = "#009E73", #green, PNW
  SW = "#E69F00", #orange, SSW
  SD = "#D55E00", #yellow, SSC
  KER = "#F0E442", #vermillion
  WMT = "#000066")#navy  

ggplot() +
  geom_sf(data = genoscape_polygons2,alpha = 0.75, aes(fill = br_node)) +
  scale_fill_manual(values = cluster_colors) +
  theme_bw()
```

```{r}
genoscape_polygons2 %>% filter(br_node=="EST") %>% write_sf("input/MigConn/WIFL/breeding_nodes/WIFL_genoscape_K1_EST.shp")
genoscape_polygons2 %>% filter(br_node=="SD") %>% write_sf("input/MigConn/WIFL/breeding_nodes/WIFL_genoscape_K2_SD.shp")
genoscape_polygons2 %>% filter(br_node=="SW") %>% write_sf("input/MigConn/WIFL/breeding_nodes/WIFL_genoscape_K3_SW.shp")
genoscape_polygons2 %>% filter(br_node=="PNW_INW") %>% write_sf("input/MigConn/WIFL/breeding_nodes/WIFL_genoscape_K4_PNW_INW.shp")

genoscape.EST <- st_read("input/MigConn/WIFL/breeding_nodes/WIFL_genoscape_K1_EST.shp") %>%   st_set_crs(4326) %>% st_transform(crs=4326)
genoscape.SD <- st_read("input/MigConn/WIFL/breeding_nodes/WIFL_genoscape_K2_SD.shp") %>% st_set_crs(4326) %>% st_transform(crs=4326)
genoscape.SW <- st_read("input/MigConn/WIFL/breeding_nodes/WIFL_genoscape_K3_SW.shp") %>% st_set_crs(4326) %>% st_transform(crs=4326)
genoscape.PNW_INW <- st_read("input/MigConn/WIFL/breeding_nodes/WIFL_genoscape_K4_PNW_INW.shp") %>% st_set_crs(4326) %>% st_transform(crs=4326)
```
```{r}
genoscape <- rbind(genoscape.EST,
                   genoscape.SW,
                   genoscape.SD,
                   genoscape.PNW_INW)
genoscape$Cluster <- c("EST", "SD","SW", "PNW_INW")
origin.names <- genoscape$Cluster
origin.names

#Nothing assigned to SD
genoscape <- rbind(genoscape.EST,
                   genoscape.SW,
                   genoscape.PNW_INW)
genoscape$Cluster <- c("EST","SW", "PNW_INW")
origin.names <- genoscape$Cluster
origin.names

```

## Nonbreeding ecoregions

```{r}
library(sf)
ecoregions<-read_sf("input/MigConn/WIFL/nonbreeding_nodes/WIFL_winterEcoregions6_simple_sf_WGS84_eBird.shp") %>% 
  filter(Region!="Northern Andes") %>% 
  sf::st_as_sf()
summary(ecoregions)

nonbr_abunds<-rast("input/MigConn/WIFL/wilfly_abundance_seasonal_nonbreeding_mean_2023.tif")

nonbr_abunds_proj <- terra::project(nonbr_abunds, crs(ecoregions))

winter_abunds <- get_vector_abunds(populations = ecoregions, abunds = nonbr_abunds_proj,pop_names=ecoregions[[6]])

target.names <- ecoregions$Region
target.names
```
```{r}
#Without NAndes because when include, nothing actually goes there
target.abund <- c(100.859306, 38.20193831, 1175.865933,11.02263654,226.1743285) #earlier abundance version

#target.abund <- c(91, 4, 625,22,268) #2023 numbers if code used above

target.abund <- target.abund / sum(target.abund)
target.abund
```

## Target points

```{r}
target.points <- read.delim ("input/MigConn/WIFL/wifl.wint128.thin_10_all.assigned_pop.REF_exclWTMT.wide_consistent.angsd.loco-pipe.txt",sep="\t") %>%
  st_as_sf(coords = c("Long", "Lat"))
target.points
```

## Get assignment in order and create matrix

```{r}
#lcwg WGSassign consistency with angsd loco-pipe pipeline
assign_mc2<-read.delim("input/MigConn/WIFL/wifl.wint128.thin_10_all.assigned_pop.REF_exclWTMT.wide_consistent.angsd.loco-pipe.txt",sep="\t")  %>% 
  mutate(SD=0) %>% 
  dplyr::select(EST,SW,PNW_INW) %>% 
  as.matrix()

```

```{r}
st_crs(target.points) <- st_crs(genoscape) <- st_crs(ecoregions)
est.psi <- estTransition(originSites = genoscape,
                          targetSites = ecoregions,
                          targetPoints = target.points,
                          originAssignment = assign_mc2,
                          isTelemetry = FALSE,
                          isRaster = FALSE,
                          isProb = TRUE,
                          captured = "target",
                          targetNames = target.names,
                          originNames = origin.names,
                          targetRelAbund = target.abund,
                          verbose = 3,
                          nSamples = 100
                          )


```

## Plot transitions
```{r}
saveRDS(est.psi,"output/MigConn/WIFL/WIFL.psi_estimates.MC.loco-pipe.exclSD0.exclNAndes.updatedMigConn.RDS")

pdf("output/MigConn/WIFL/WIFL.estTrans_plot.MC.loco-pipe.pdf")
par(mar = c(2,3,0,4),xpd = TRUE)
plot(est.psi,legend ="topright",cex=1)
dev.off()
```
## Print mean transition probaibilities
```{r}
est.psi$psi$mean
```

## Uses the transition probabilities to estimate (node-specific) network migratory connectivty strength, or what we call NMCpop in MigConnectivity.

This is the spread on the wintering ground

```{r}
est.psi<-readRDS("output/MigConn/WIFL/WIFL.psi_estimates.MC.loco-pipe.exclSD0.exclNAndes.updatedMigConn.RDS")

calcNMC(est.psi$psi$mean)

estNMC(
  est.psi,
  originNames = NULL,
  targetNames = NULL,
  originSites = NULL,
  targetSites = NULL,
  nSamples = 1000,
  verbose = 0,
  alpha = 0.05,
  returnAllInput = TRUE
)
```
## Didn't do this, but code included

We then use the resulting object to estimate the strength of migratory connectivity. We first need some additional data, the distances between originSites and targetSites.
```{r}
# calculate distance between originSites
originCentersWIFL <- st_centroid(genoscape)

originCentersWIFL <- st_transform(originCentersWIFL, 4326)
originDistWIFL <- distFromPos(st_coordinates(originCentersWIFL$geometry))

# calculate distance between targetSites 
targetCentersWIFL <- st_centroid(ecoregions)

targetCentersWIFL <- st_transform(targetCentersWIFL, 4326)
targetDistWIFL <- distFromPos(st_coordinates(targetCentersWIFL$geometry))
```

# Estimate the strength of migratory connectivity
```{r}
origin.abund <- c(40601,489,41951)
origin.abund <- origin.abund / sum(origin.abund)
origin.abund

WIFLmc <- estStrength(originDist = originDistWIFL,
                      targetDist = targetDistWIFL,
                      originRelAbund =  origin.abund,
                      psi = est.psi,
                      nSamples = 5000,
                      verbose = 0)
WIFLmc 
```
